**HACK** - Fix: MissingMethodException using PrivateKeys
Using Private Key access results in an exception in VS2022 when MAUI workloads are installed because they force the Neon.SSH component to use the newest version (loaded by VS). Updates to Neon.SSH are beyond our control, so hack a solution:  Clone Neon.SSH in an adjacent dir to this project, reference the Neon components specifically, and force the update of SSH.NET that's required to avoid the exception. The update forces a single line change for a nullable at present. Otherwise, it's just a recompile to pick up on the introduction of an Interface.