Include PlatformTarget in IsRaspberryCompatable
If the PlatformTarget is ARM32/64 then assume this can be executed on a Pi.  This is sorta a stop gap because it is expected that the project sets the RuntimeIdentifier to "linux-arm64" but the property isn't included in the active config (or anywhere I could find) so use this for now.