Add code to determine what version of VS is being used by the debugger.
This code will manage the version of VS that the extension is currently being used in.
It helps determine what debugger should be installed and where it is installed on the Pi.  MS docs suggest that the version can be selected but don't really provide info on the valid values. Also, while vs2022 works, vs2026 currently does not. So just go with latest for now.